# [8_11] optimize performance of as_hexadecimal